Changed some behaviour of RoomServiceImpl class, also created new UserNotFoundException

+++ src/main/java/api/service/RoomServiceImpl.java
@@ -1,6 +1,7 @@
 package api.service;
 
 import api.entity.Room;
+import api.util.exception.RoomNotFoundException;
 import api.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -18,25 +19,44 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     @Transactional
+    // TODO: Make Sorting
     public List<Room> getAllRooms() {
         return roomRepository.findAll();
     }
 
     @Override
     @Transactional
-    public Room getRoomById(int id) {
-        return roomRepository.findById(id).orElse(null);
+    public Room getRoomById(int id) throws RoomNotFoundException {
+        if (roomRepository.findById(id).isPresent()) {
+            return roomRepository.findById(id).get();
+        } else {
+            throw new RoomNotFoundException("Room with this id not found");
+        }
     }
 
     @Override
     @Transactional
-    public Room saveOrUpdateRoom(Room room) {
+    public Room saveRoom(Room room) {
         return roomRepository.save(room);
     }
 
     @Override
     @Transactional
+    public Room updateRoom(Room room) {
+        if(roomRepository.existsById(room.getId())) {
+            return roomRepository.save(room);
+        } else {
+            throw new RoomNotFoundException("Room with this id not found");
+        }
+    }
+
+    @Override
+    @Transactional
     public void deleteRoom(int id) {
-        roomRepository.deleteById(id);
+        if(roomRepository.existsById(id)) {
+            roomRepository.deleteById(id);
+        } else {
+            throw new RoomNotFoundException("Room with this id not found");
+        }
     }
 }
